fix: Small alignments on local-notifications
